:sparkles: Initialize repo
The repo is already initialized with a lot of Docker and Rails configuration. We already have all of the models, controllers, helpers, seeds, abilities, etc..., because this is an old project being migrated to Rails 7.

The landing page is also functional, but we don't have the login and dashboard pages yet.